+++ packages/documentation/src/stories/components/forms/input/input.snapshot.stories.ts
@@ -1,10 +1,12 @@
 import type { Args, StoryContext, StoryObj } from '@storybook/web-components';
 import meta from './input.stories';
 import { html } from 'lit';
-import { getCombinations, COMBINATIONS } from '../../../../utils/inputComponentsGetCombinations';
+import { COMBINATIONS, getCombinations } from '../../../../utils/inputComponentsGetCombinations';
+
+const { id, ...metaWithoutId } = meta;
 
 export default {
-  ...meta,
+  ...metaWithoutId,
   title: 'Snapshots',
   render: renderInputSnapshot,
 };
@@ -19,6 +21,10 @@ function renderInputSnapshot(_args: Args, context: StoryContext) {
     {
       label: `Label - with Value`,
       value: 'Lorem Ipsum',
+    },
+    {
+      label: `Label - Floating label`,
+      floatingLabel: true,
     },
   ];
   return html`
@@ -61,9 +67,7 @@ function renderInputSnapshot(_args: Args, context: StoryContext) {
               )
               .map(
                 (args: Args) =>
-                  html`
-                    <div>${meta.render?.({ ...context.args, ...args }, context)}</div>
-                  `,
+                  html` <div>${meta.render?.({ ...context.args, ...args }, context)}</div> `,
               )}
           </div>
         `,
@@ -107,10 +111,5 @@ export const Inputweek: Story = {
 export const Inputtime: Story = {
   args: {
     type: 'time',
-  },
-};
-export const Inputcolor: Story = {
-  args: {
-    type: 'color',
   },
 };