+++ packages/documentation/src/stories/components/icons/getting-started.stories.ts
@@ -2,6 +2,7 @@ import { Meta, StoryObj } from '@storybook/web-components';
 import { BADGE } from '../../../../.storybook/constants';
 
 const meta: Meta = {
+  id: '40ed323b-9c1a-42ab-91ed-15f97f214608',
   title: 'Components/Icons/Getting Started',
   parameters: {
     badges: [BADGE.NEEDS_REVISION],