+++ packages/documentation/src/stories/getting-started/styles/styles.stories.ts
@@ -2,6 +2,7 @@ import { Meta, StoryObj } from '@storybook/web-components';
 import { BADGE } from '../../../../.storybook/constants';
 
 const meta: Meta = {
+  id: 'e53e2de8-0bbf-4f70-babc-074c5466f700',
   title: 'Getting Started/Styles',
   parameters: {
     badges: [BADGE.NEEDS_REVISION],